move stagehand toolkit and tests to toolkits dir
